+++ .babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "env"
+  ],
+  "plugins": [
+    ["transform-react-jsx", { "pragma": "h" }]
+  ]
+}
