--- background.js
@@ -1,2 +1,0 @@
-var app = new App();
-app.constructor();
