+++ src/style.css
@@ -1,0 +1,93 @@
+html, body, div, span,
+h1, h2, h3, h4, h5, h6, p, blockquote, pre,
+a, footer {
+	margin: 0;
+	padding: 0;
+	border: 0;
+	font-size: 100%;
+	font: inherit;
+	vertical-align: baseline;
+}
+body {
+	line-height: 1;
+}
+blockquote, q {
+	quotes: none;
+}
+* {box-sizing: border-box;}
+html, body {
+  font-family: 'Spectral', serif;
+  color:white;
+  -webkit-font-smoothing: antialiased;
+}
+#view {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+}
+#app {
+  height: 100vh;
+}
+body.dark {
+  background-color: #2c2c2c;
+  color:white;
+}
+body.light {
+  background-color: #eee;
+  color:#555;
+}
+a {
+	color:inherit;
+	text-decoration: none;
+}
+@keyframes fadeIn {
+  from {opacity: 0}
+  to {opacity: 1}
+}
+#showerthought {
+  opacity: 0;
+  animation: 1s fadeIn forwards;
+  padding: 2em;
+  position: relative;
+}
+#showerthought blockquote {
+  font-size: 1.8em;
+  line-height: 130%;
+	margin-bottom: 1em;
+	text-indent: -1ch;
+}
+#showerthought span.quote {
+	opacity: 0.5;
+}
+#showerthought p {
+  font-size: 1.3em;
+	opacity: 0.6;
+	font-style: italic;
+}
+footer {
+  position: absolute;
+  bottom: 0;
+  padding: 1em;
+  opacity: 0.4;
+  transition: opacity 0.3s ease;
+  width: 100%;
+}
+footer:hover {
+  opacity: 1;
+}
+.btn {
+  border: none;
+  outline: none;
+  background-color: rgba(0, 0, 0, 0.2);
+  border-radius: 0.2em;
+  padding: 8px 12px;
+  text-transform: uppercase;
+  font-weight: bold;
+  color:white;
+  margin-right: 0.3em;
+  cursor: pointer;
+}
+.btn:hover {
+  background-color: #3498db;
+}
