+++ src/tab.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>New Tab</title>
+  <link rel="stylesheet" href="./style.css">
+  <link href="https://fonts.googleapis.com/css?family=Spectral:400,500" rel="stylesheet">
+</head>
+<body class="light">
+  <div id="view"></div>
+  <footer>
+    <button class="btn" id="theme-toggle">Toggle Theme</button>
+  </footer>
+  <script src="./app.js"></script>
+</body>
+</html>
